refactor: SettingEvent 모델 분리
