+++ android/app/src/main/java/com/yagubogu/ui/setting/SettingAccountScreen.kt
@@ -22,6 +22,7 @@ import com.yagubogu.ui.setting.component.SettingButtonGroup
 import com.yagubogu.ui.setting.component.SettingEventHandler
 import com.yagubogu.ui.setting.component.dialog.SettingDialog
 import com.yagubogu.ui.setting.component.model.SettingDialogEvent
+import com.yagubogu.ui.setting.component.model.SettingEvent
 import com.yagubogu.ui.theme.Gray050
 
 @Composable
